fbService structure
